+++ VpnHood.Client/VpnHoodConnect.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.Extensions.Logging;
 using VpnHood.Client.Device;
 using VpnHood.Common;
 using VpnHood.Common.Logging;
@@ -107,11 +105,11 @@ public class VpnHoodConnect : IAsyncDisposable
         return DisposeAsync(true);
     }
 
-    public async ValueTask DisposeAsync(bool waitForBye)
+    public ValueTask DisposeAsync(bool waitForBye)
     {
         lock (_disposeLock)
             _disposeTask ??= DisposeAsyncCore(waitForBye);
-        await _disposeTask.Value;
+        return _disposeTask.Value;
     }
 
     private async ValueTask DisposeAsyncCore(bool waitForBye)