+++ VpnHood.Server.App.Net/AppSettings.cs
@@ -1,4 +1,4 @@
-﻿using System;
+﻿using System.Text.Json.Serialization;
 using VpnHood.Server.Access.Configurations;
 using VpnHood.Server.Access.Managers.File;
 using VpnHood.Server.Access.Managers.Http;
@@ -8,16 +8,26 @@ namespace VpnHood.Server.App;
 
 public class AppSettings
 {
-    [Obsolete ("Use HttpAccessManager")]
-    public HttpAccessManagerOptions? RestAccessServer { get=> HttpAccessManager; set => HttpAccessManager = value;}
-    [Obsolete ("Use HttpAccessManager")]
-    public HttpAccessManagerOptions? HttpAccessServer { get=> HttpAccessManager; set => HttpAccessManager = value;}
+    [Obsolete("Use HttpAccessManager")]
+    public HttpAccessManagerOptions? RestAccessServer { set => HttpAccessManager = value; }
+    [Obsolete("Use HttpAccessManager")]
+    public HttpAccessManagerOptions? HttpAccessServer { set => HttpAccessManager = value; }
     [Obsolete("Use FileAccessManager")]
-    public FileAccessManagerOptions? FileAccessServer { get => FileAccessManager; set => FileAccessManager = value; }
+    public FileAccessManagerOptions? FileAccessServer { set => FileAccessManager = value; }
+    [Obsolete("Use ManagementSecret")]
+    public string? Secret { set => ManagementSecret = value; }
 
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
     public HttpAccessManagerOptions? HttpAccessManager { get; set; }
-    public FileAccessManagerOptions? FileAccessManager { get; set; } = new();
+    
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+    public FileAccessManagerOptions? FileAccessManager { get; set; }
+    
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
     public ServerConfig? ServerConfig { get; set; }
+
     public bool AllowAnonymousTracker { get; set; } = true;
     public bool IsDiagnoseMode { get; set; }
+
+    public string? ManagementSecret { get; set; }
 }