+++ VpnHood.Server.App.Net/FileAccessManagerCommand.cs
@@ -1,9 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
+﻿using System.Net;
 using System.Text.Json;
-using System.Threading.Tasks;
 using McMaster.Extensions.CommandLineUtils;
 using Microsoft.Extensions.Logging;
 using VpnHood.Common;
@@ -113,7 +109,7 @@ public class FileAccessManagerCommand
                     item.Port = bestEp?.Port ?? 443;
                 }
 
-                // throw error if could not find any public endpoint
+                // throw error if there is no any public endpoint
                 if (publicEndPoints.Length == 0)
                 {
                     VhLogger.Instance.LogError(