+++ VpnHood.Tunneling/Channels/Streams/BinaryStream.cs
@@ -1,9 +1,5 @@
 ﻿using Microsoft.Extensions.Logging;
-using System;
-using System.IO;
 using System.Net.Sockets;
-using System.Threading;
-using System.Threading.Tasks;
 using VpnHood.Common.Logging;
 using VpnHood.Common.Utils;