+++ VpnHood.Tunneling/DatagramMessaging/DatagramMessageHandler.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.IO;
-using System.Net;
+﻿using System.Net;
 using PacketDotNet;
 
 namespace VpnHood.Tunneling.DatagramMessaging;